// frontend/src/api/model/diagnosisRole.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Qué lugar ocupa el diagnóstico en la internación.
 */
export type DiagnosisRole = typeof DiagnosisRole[keyof typeof DiagnosisRole];


export const DiagnosisRole = {
  PRINCIPAL: 'PRINCIPAL',
  SECONDARY: 'SECONDARY',
  COMORBIDITY: 'COMORBIDITY',
  COMPLICATION: 'COMPLICATION',
} as const;