// frontend/src/api/model/dischargeDestination.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type DischargeDestination = typeof DischargeDestination[keyof typeof DischargeDestination];


export const DischargeDestination = {
  HOME: 'HOME',
  OTHER_FACILITY: 'OTHER_FACILITY',
  HOME_CARE: 'HOME_CARE',
  REHABILITATION: 'REHABILITATION',
  OTHER: 'OTHER',
} as const;