// frontend/src/api/model/dischargePlanCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { DischargeDestination } from './dischargeDestination';
import type { DischargePlanStatus } from './dischargePlanStatus';

export interface DischargePlanCreate {
  planned_date?: string | null;
  destination?: DischargeDestination;
  requires_transport?: boolean;
  requires_home_care?: boolean;
  status?: DischargePlanStatus;
  created_by?: string | null;
  notes?: string | null;
}