// frontend/src/api/model/dischargeType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type DischargeType = typeof DischargeType[keyof typeof DischargeType];


export const DischargeType = {
  MEDICAL: 'MEDICAL',
  VOLUNTARY: 'VOLUNTARY',
  TRANSFER: 'TRANSFER',
  DECEASED: 'DECEASED',
  ABSCONDED: 'ABSCONDED',
  OTHER: 'OTHER',
} as const;